requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.2
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
idna==3.10
numpy==2.2.3
pandas==2.2.3
proto-plus==1.26.0
protobuf
psycopg2-binary
py4j==0.10.9.7
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyspark==3.5.5
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
rsa==4.9
setuptools==75.8.2
six==1.17.0
SQLAlchemy
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
folium
matplotlib
papermill
pylint

astro-run-dag # This package is needed for the astro run command. It will be removed before a deploy